Move layout manager model to separate files
