+++ python/gui/auto_generated/layout/qgslayoutcombobox.sip.in
@@ -8,6 +8,7 @@
 
 
 
+
 class QgsLayoutComboBox : QComboBox
 {
 %Docstring(signature="appended")
